+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 [Back homepage](https://github.com/cthadeusantos/makemake)
 # Changes
 
+## 0.1.1 (2025/01/02)
+* Fix wrong number for project code at project app
+* Implemented automatic filename for new documents
+* Replace check cpf and cnpj numbers functions (Brazil numbers)
+* Build template tags and filters for cpf/cnpj (Brazil numbers format)
+
 ## 0.1.0 (2024/12/31)
 * First release